backend-fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "1.0.0"
description = "FastAPI backend with MongoDB and OpenAI integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "motor>=3.3.2",
    "openai>=1.3.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.1",
]
requires-python = ">=3.8.1"
readme = "README.md"
license = {text = "MIT"}
[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
test = [
    "pytest>=7.4.3",
    "httpx>=0.25.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"